#7 Displays top 10 emails in a basic table format. Need to implement a component library for styling
