+++ ec_email_client/src/App.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import logo from "./logo.svg";
 import "./App.css";
+import InboxPage from "./pages/InboxPage";
 
 function App() {
     // Client ID and API key from the Developer Console
@@ -17,6 +18,8 @@ function App() {
     // included, separated by spaces.
     var SCOPES =
         "https://www.googleapis.com/auth/gmail.readonly https://www.googleapis.com/auth/gmail.send https://www.googleapis.com/auth/gmail.compose https://www.googleapis.com/auth/gmail.modify";
+
+    const [profileInfo, setProfileInfo] = useState(null);
 
     function handleClientLoad() {
         console.log("About to loadwindow.gapi");
@@ -69,6 +72,7 @@ function App() {
         console.log("Sign in status: " + isSignedIn);
         if (isSignedIn) {
             console.log("Profile Info: ", await getProfileInfo("me"));
+            setProfileInfo(await getProfileInfo("me"));
             // let messageIds = await getMessagesIds();
             // console.log(messageIds);
             // console.log(await getMessage(messageIds[0].id));
@@ -200,152 +204,9 @@ function App() {
             <button id="signout_button" style={{ display: "none" }}>
                 Sign Out
             </button>
-            <pre id="content" style={{ "white-space": "pre-wrap" }}></pre>
-            <InboxPage />
+            {/* Once profile is loaded, we know GAPI is loaded, so get emails */}
+            {profileInfo && <InboxPage />}
         </div>
-    );
-}
-
-function InboxPage(props) {
-    async function getMessagesIds(userId) {
-        // To get userId of logged in user, give "me"
-        if (userId === undefined) userId = "me";
-
-        return new Promise((resolve, reject) => {
-            window.gapi.client.gmail.users.messages
-                .list({
-                    userId: userId,
-                })
-                .then(function (response) {
-                    resolve(response.result.messages);
-                });
-        });
-    }
-
-    async function getMessage(messageId) {
-        // https://developers.google.com/gmail/api/reference/rest/v1/users.messages.attachments#MessagePartBody
-        // console.log("Getting specific message for " + messageId);
-
-        let message = {
-            bodyHTML: null,
-            bodyText: null,
-            from: null,
-            to: null,
-            subject: null,
-            headers: {},
-        };
-
-        return new Promise((resolve, reject) => {
-            window.gapi.client.gmail.users.messages
-                .get({
-                    userId: "me",
-                    id: messageId,
-                })
-                .then(function (response) {
-                    // console.log(response.result);
-
-                    // If it has two payload parts
-                    if (
-                        !!response.result.payload.parts &&
-                        response.result.payload.parts.length > 1
-                    ) {
-                        message.bodyText = decodeBase64HTML(
-                            response.result.payload.parts[0].body.data
-                        );
-                        message.bodyHTML = decodeBase64HTML(
-                            response.result.payload.parts[1].body.data
-                        );
-                    } else if (
-                        !!response.result.payload.body &&
-                        !!response.result.payload.body.data
-                    ) {
-                        message.bodyText = decodeBase64HTML(
-                            response.result.payload.body.data
-                        );
-                    } else {
-                        console.log(
-                            "Failed getting message body for:",
-                            response.result
-                        );
-                    }
-
-                    // Gets all headers, turns in to dict
-                    let headers = {};
-                    response.result.payload.headers.forEach((header) => {
-                        headers[header.name] = header.value;
-                        if (header.name === "Subject") {
-                            message.subject = header.value;
-                        }
-                        if (header.name === "To") {
-                            message.to = header.value;
-                        }
-                        if (header.name === "From") {
-                            message.from = header.value;
-                        }
-                    });
-                    message.headers = headers;
-
-                    resolve(message);
-                });
-        });
-    }
-
-    async function getAllMessages(userId) {
-        // To get userId of logged in user, give "me"
-        if (userId === undefined) userId = "me";
-        const numOfMessagesToGet = 10;
-
-        let messageIds = await getMessagesIds(userId);
-        let messages = [];
-        for (var i = 0; i < numOfMessagesToGet; i++) {
-            try {
-                messages.push(await getMessage(messageIds[i].id));
-            } catch {
-                console.log("Failed on ", messageIds[i].id);
-            }
-        }
-
-        return messages;
-    }
-
-    function decodeBase64(data) {
-        return atob(data);
-    }
-
-    function decodeBase64HTML(data) {
-        // Replace non-url compatible chars with base64 standard chars
-        let input = data.replace(/-/g, "+").replace(/_/g, "/");
-
-        // Pad out with standard base64 required padding characters
-        var pad = input.length % 4;
-        if (pad) {
-            if (pad === 1) {
-                throw new Error(
-                    "InvalidLengthError: Input base64url string is the wrong length to determine padding"
-                );
-            }
-            input += new Array(5 - pad).join("=");
-        }
-
-        return decodeBase64(input);
-    }
-
-    const [emails, setEmails] = useState([]);
-
-    return (
-        <table>
-            <InboxEmailLine />
-        </table>
-    );
-}
-
-function InboxEmailLine(props) {
-    return (
-        <tr>
-            <td>FromName</td>
-            <td>Subject - </td>
-            <td>Snippet of Message</td>
-        </tr>
     );
 }
