+++ ec_email_client/src/pages/InboxPage.css
@@ -1,0 +1,3 @@
+table {
+    width: 100%;
+}
